step-42: fix compilation error
git-svn-id: https://svn.dealii.org/trunk@25561 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/examples/step-42/step-42.cc
@@ -1177,7 +1177,7 @@ void Step4<dim>::solve_newton ()
       double a = 0;
       for (unsigned int i=0; (i<10)&&(!damped); i++)
 	{
-	  a=pow(0.5,i);
+	  a=pow(0.5, static_cast<double>(i));
 	  old_solution = tmp_vector;
 	  old_solution.sadd(1-a,a, distributed_solution);
